# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: AI QA Agent API
  version: 1.0.0
  description: |
    Backend for the AI QA Agent platform: projects, runs, test cases, runtimes
    and reports.

    Every request that touches organization data carries `X-Org-ID`, which is
    the only source of the active organization (ADR-006/ADR-007). A missing,
    unparsable, unknown or non-member organization is `403`; a resource id
    belonging to another organization is `404`, because the org-scoped query
    does not find it.
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/runs:
    post:
      tags: [runs]
      summary: Start a run
      description: |
        `runtimeId` is optional. When it is absent the scheduler assigns the run
        to any online runtime in the organization; when no runtime is available
        the run stays `queued` rather than failing.

        Send `Idempotency-Key` to make a retry safe: a replay returns the
        original run with `200` instead of starting a second browser session
        against the customer's application.
      parameters:
        - $ref: '#/components/parameters/OrgHeader'
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, mode]
              properties:
                projectId: { type: string, format: uuid }
                runtimeId: { type: [string, 'null'], format: uuid }
                mode: { $ref: '#/components/schemas/RunMode' }
                testCaseIds:
                  type: array
                  description: Explicit selection for an execute or full run. Absent means every approved case.
                  items: { type: string, format: uuid }
      responses:
        '200':
          description: An Idempotency-Key replayed an earlier request; this is the run it created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '201':
          description: Queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    get:
      tags: [runs]
      summary: List runs, newest first
      parameters:
        - $ref: '#/components/parameters/OrgHeader'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: projectId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items: { $ref: '#/components/schemas/Run' }
                  total: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/runs/{runID}:
    get:
      tags: [runs]
      summary: Get a run
      parameters:
        - $ref: '#/components/parameters/OrgHeader'
        - $ref: '#/components/parameters/RunPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/runs/{runID}/cancel:
    post:
      tags: [runs]
      summary: Cancel a run
      description: |
        Idempotent: cancelling an already-cancelled run succeeds. Cancelling a
        run that finished on its own is a 409, because that outcome is no longer
        reachable. The assigned daemon receives a `run.cancel` frame.
      parameters:
        - $ref: '#/components/parameters/OrgHeader'
        - $ref: '#/components/parameters/RunPath'
      responses:
        '200':
          description: Cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/runs/{runID}/events:
    get:
      tags: [runs]
      summary: A run's event stream
      description: |
        The polling counterpart of the WebSocket. Events are numbered from 0 per
        run; pass the previous page's `nextSince` to resume. Delivery from the
        daemon is at-least-once and deduplicated on `(runId, seq)`, so a
        sequence number appears exactly once here.
      parameters:
        - $ref: '#/components/parameters/OrgHeader'
        - $ref: '#/components/parameters/RunPath'
        - $ref: '#/components/parameters/Limit'
        - name: since
          in: query
          description: Return events after this sequence number. Omit or use -1 for the beginning.
          schema: { type: integer, format: int64, default: -1 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items: { $ref: '#/components/schemas/RunEvent' }
                  nextSince: { type: integer, format: int64 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/runs/{runID}/artifacts/presign:
    post:
      tags: [runs]
      summary: Mint an upload URL for one artifact
      description: |
        Daemon-facing; authenticated with the runtime bearer token, not a
        session. The returned URL is a presigned S3 PUT for exactly the key
        asked for, and the key must be inside this run's own
        `orgs/{orgId}/runs/{YYYY-MM-DD}/{runId}/` prefix — a key outside it is a
        403. See the package doc of `server/internal/artifact` for why the grant
        is a minting endpoint rather than one prefix-wide signature.
      security:
        - runtimeToken: []
      parameters:
        - $ref: '#/components/parameters/RunPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key: { type: string, maxLength: 1024 }
                contentType: { type: string }
      responses:
        '201':
          description: Signed
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string, format: uri }
                  key: { type: string }
                  method: { type: string, const: PUT }
                  expiresAt: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: qa_session
    runtimeToken:
      type: http
      scheme: bearer
      description: The long-lived runtime token issued by /api/v1/runtimes/redeem.
  parameters:
    OrgHeader:
      name: X-Org-ID
      in: header
      required: true
      description: The active organization. The only source of an org id on a request.
      schema: { type: string, format: uuid }
...
    RunPath:
      name: runID
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Not signed in, or the credential is not valid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Signed in, but not allowed — including every X-Org-ID failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such row, or it belongs to another organization.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: The request is well formed but the current state refuses it.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: The JSON parsed; the values are not acceptable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: A dependency this route needs is not configured or not reachable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, validation_failed, unauthorized, forbidden, not_found,
                     method_not_allowed, conflict, unsupported_media_type, payload_too_large,
                     rate_limited, internal, unavailable, timeout]
            message: { type: string }
            details: { type: object }
...
    RunMode:
      type: string
      enum: [discover, plan, execute, full]
    RunStatus:
      type: string
      enum: [queued, assigned, running, passed, failed, cancelled, error]
    Run:
      type: object
      required: [id, projectId, mode, status, phase, counters, attempts, createdAt]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        runtimeId: { type: [string, 'null'], format: uuid }
        mode: { $ref: '#/components/schemas/RunMode' }
        status: { $ref: '#/components/schemas/RunStatus' }
        phase:
          type: string
          description: Free-form progress label. Not an enum, because the pipeline grows steps faster than a migration can follow.
        counters:
          type: object
          properties:
            total: { type: integer }
            passed: { type: integer }
            failed: { type: integer }
            skipped: { type: integer }
            errored: { type: integer }
        error:
          type: object
          description: A domain error, never a driver message.
          properties:
            code: { type: string }
            message: { type: string }
        attempts: { type: integer }
        createdAt: { type: string, format: date-time }
        assignedAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
    RunEvent:
      type: object
      required: [seq, phase, level, code, message, data, ts]
      properties:
        seq: { type: integer, format: int64 }
        phase: { type: string }
        level: { type: string, enum: [debug, info, warn, error] }
        code:
          type: string
          description: Machine-readable event code. The UI switches on this, never on message.
        message: { type: string }
        data: { type: object }
        ts: { type: string, format: date-time }
...